Final Version 27/08/2021

+++ app/src/main/java/gcubeit/com/enerwireproduccionv12/data/database/entity/DbConversion.kt
@@ -2,17 +2,14 @@ package gcubeit.com.enerwireproduccionv12.data.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.google.gson.annotations.SerializedName
-import gcubeit.com.enerwireproduccionv12.data.network.response.color.Color
 import gcubeit.com.enerwireproduccionv12.data.network.response.conversion.Conversion
-import gcubeit.com.enerwireproduccionv12.data.network.response.product.Product
 
 @Entity(tableName = "conversion_table")
 data class DbConversion(
     @PrimaryKey(autoGenerate = false)
     val id: Int,
     val description: String,
-    val factor: Double,
+    val factor: Float,
     val type: String
 )
